move useGetProjectMetadata to own file
